requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
aiofiles==23.2.1